pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "fastapi[standard]>=0.115.13",
    "pydantic-settings>=2.10.1",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.0",
]
